# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Life Kline Commercial API
  version: 1.0.0
  description: 商用 iOS AI 应用后端 API 契约。
servers:
  - url: http://127.0.0.1:3000
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /api/session:
    get:
      summary: Restore current session
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: Session bundle or unauthenticated state.
  /api/auth/passwordless:
    post:
      summary: Create or restore a passwordless session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthPayload"
      responses:
        "200":
          description: Login success.
        "422":
          description: Validation failed.
...
  /api/auth/logout:
    post:
      summary: Revoke current session
      responses:
        "200":
          description: Logged out.
        "401":
          description: Authentication required.
  /api/me:
    get:
      summary: Get current user bundle
      responses:
        "200":
          description: Current user bundle.
        "401":
          description: Authentication required.
...
  /api/profile:
    post:
      summary: Save user birth profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [profile]
              properties:
                profile:
                  $ref: "#/components/schemas/UserProfile"
                bazi:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: Profile saved.
        "401":
          description: Authentication required.
        "422":
          description: Validation failed.
...
  /api/membership/checkout:
    post:
      summary: Create a membership order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutPayload"
      responses:
        "200":
          description: Order created or paid in mock mode.
        "401":
          description: Authentication required.
        "422":
          description: Validation failed.
  /api/membership/confirm:
    post:
      summary: Confirm a live payment receipt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [orderId]
              properties:
                orderId:
                  type: string
                  format: uuid
                providerOrderId:
                  type: string
                  maxLength: 160
                receipt:
                  type: string
                  maxLength: 4096
      responses:
        "200":
          description: Payment confirmed.
        "401":
          description: Authentication required.
        "422":
          description: Validation failed.
...
  /api/settings:
    post:
      summary: Save user settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [settings]
              properties:
                settings:
                  $ref: "#/components/schemas/UserSettings"
      responses:
        "200":
          description: Settings saved.
  /api/bindings:
    post:
      summary: Save account bindings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bindings]
              properties:
                bindings:
                  $ref: "#/components/schemas/AccountBindings"
      responses:
        "200":
          description: Bindings saved.
  /api/share-count:
    post:
      summary: Save share unlock count
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [shareCount]
              properties:
                shareCount:
                  type: integer
                  minimum: 0
                  maximum: 1000
      responses:
        "200":
          description: Share count saved.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: life_session
  schemas:
    AuthPayload:
      type: object
      required: [provider, clientInstallationId]
      properties:
        provider:
          type: string
          enum: [wechat, apple, google, phone, email, guest]
        clientInstallationId:
          type: string
          maxLength: 80
        providerSubject:
          type: string
          maxLength: 160
        displayName:
          type: string
          maxLength: 80
        email:
          type: string
          format: email
        phone:
          type: string
          maxLength: 32
...
    UserProfile:
      type: object
      required: [name, gender, birthDate, birthTime, birthPlace]
      properties:
        name:
          type: string
          maxLength: 80
        gender:
          type: string
          enum: [male, female]
        birthDate:
          type: string
          pattern: "^\\d{4}-\\d{2}-\\d{2}$"
        birthTime:
          type: string
          pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"
        birthPlace:
          type: string
          maxLength: 160
...
    CheckoutPayload:
      type: object
      required: [plan, provider, amountCents]
      properties:
        plan:
          type: string
          enum: [monthly, lifetime]
        provider:
          type: string
          enum: [wechat, alipay, apple]
        amountCents:
          type: integer
          minimum: 100
          maximum: 999900
    UserSettings:
      type: object
      properties:
        notifications:
          type: boolean
        language:
          type: string
          maxLength: 40
    AccountBindings:
      type: object
      properties:
        phone:
          type:
            - string
            - "null"
          maxLength: 32
        wechat:
          type: boolean